fix: close portal and caller attribution review gaps
