+++ Sources/GhosttyTerminalView.swift
@@ -13567,6 +13567,11 @@ struct GhosttyTerminalView: NSViewRepresentable {
         coordinator.lastBoundHostId = nil
         coordinator.portalReconciliationScheduler.cancel()
         let hostedView = coordinator.hostedView
+        let host = nsView as? HostContainerView
+        let wasBoundToDismantledHost: Bool = {
+            guard let host, let hostedView else { return false }
+            return TerminalWindowPortalRegistry.hasEntry(for: hostedView, boundTo: host)
+        }()
 #if DEBUG
         if let hostedView {
             if let snapshot = AppDelegate.shared?.tabManager?.debugCurrentWorkspaceSwitchSnapshot() {
@@ -13585,7 +13590,14 @@ struct GhosttyTerminalView: NSViewRepresentable {
         }
 #endif
 
-        if let host = nsView as? HostContainerView {
+        // Only the host that is still bound to this surface may clear the
+        // shared ring. Do this before preparing a replacement so a synchronous
+        // hand-off cannot let the old teardown hide the new owner's ring.
+        if wasBoundToDismantledHost {
+            hostedView?.setNotificationRing(visible: false)
+        }
+
+        if let host {
             host.onDidMoveToWindow = nil
             host.onGeometryChanged = nil
             // The owner's vacate path drops its own wake-up; a candidate that
@@ -13603,7 +13615,6 @@ struct GhosttyTerminalView: NSViewRepresentable {
 
         // Preserve the portal lease across transient rebuilds, but reset the
         // surface-local ring; the next reconciliation reapplies current state.
-        hostedView?.setNotificationRing(visible: false)
         hostedView?.setFocusHandler(nil)
         hostedView?.setTriggerFlashHandler(nil)
         hostedView?.setDropZoneOverlay(zone: nil)
