+++ cmuxTests/TerminalNotificationSocketActionTests.swift
@@ -1,4 +1,4 @@
-import XCTest
+@preconcurrency import XCTest
 import AppKit
 import Darwin
 #if canImport(cmux_DEV)
@@ -7,6 +7,8 @@ import Darwin
 @testable import cmux
 #endif
 
+// This existing socket suite remains on XCTest because its tests share the
+// XCTestCase setup/teardown and socket fixture lifecycle below.
 @MainActor
 final class TerminalNotificationSocketActionTests: XCTestCase {
     override func setUp() {
@@ -234,7 +236,7 @@ final class TerminalNotificationSocketActionTests: XCTestCase {
         XCTAssertEqual(fixture.notification(openable.id)?.isRead, true)
     }
 
-    private struct SocketFixture {
+    struct SocketFixture {
         let socketPath: String
         let store: TerminalNotificationStore
         let appDelegate: AppDelegate
@@ -274,7 +276,7 @@ final class TerminalNotificationSocketActionTests: XCTestCase {
         }
     }
 
-    private func makeSocketFixture(name: String, includeWindow: Bool = false) throws -> SocketFixture {
+    func makeSocketFixture(name: String, includeWindow: Bool = false) throws -> SocketFixture {
         let socketPath = makeSocketPath(name)
         let store = TerminalNotificationStore.shared
         let previousShared = AppDelegate.shared
@@ -384,7 +386,7 @@ final class TerminalNotificationSocketActionTests: XCTestCase {
         )
     }
 
-    private func sendV2RequestAsync(
+    func sendV2RequestAsync(
         method: String,
         params: [String: Any] = [:],
         to socketPath: String
